requirements.txt
fastapi
uvicorn
playwright
requests
beautifulsoup4
pydantic
httpx
python-dotenv